stats component integration fix
